requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
idna==3.4
lxml==4.9.2
packaging==23.0
parsel==1.7.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
selectorlib==0.16.0
urllib3==1.26.15
w3lib==2.1.1
